test: isolate effective binding label coverage
